Added base classes
Added a base web crawler class, as well as an extending class that crawls a specific website. Also added a configuration class, so you can configure crawl requests using a configuration file. There is also a basic debugger that prints warnings, errors, and debugging messages with the calling function's name